+++ agent/mibgroup/utilities/execute.c
@@ -354,7 +354,7 @@ run_exec_command( char *command, char *input,
              */
             count = read(opipe[0], &cache_ptr[offset], cache_size);
             DEBUGMSGTL(("verbose:run:exec",
-                        "    read %d bytes\n", count));
+                        "    read %d bytes\n", (int)count));
             if (0 == count) {
                 int rc;
                 /*
@@ -389,7 +389,7 @@ run_exec_command( char *command, char *input,
                         break;
                     }
                     DEBUGMSGTL(("verbose:run:exec",
-                                "    %d left in buffer\n", cache_size));
+                                "    %d left in buffer\n", (int)cache_size));
                 }
             }
             else if ((count == -1) && (EAGAIN != errno)) {
